+++ src/Libraries/RevitNodes/GeometryObjects/GeometryObjectSelector.cs
@@ -67,6 +67,8 @@ namespace Revit.GeometryObjects
         /// <returns></returns>
         private static bool RequiresTransform(Autodesk.Revit.DB.FamilyInstance familyInstance)
         {
+            if (familyInstance.Category.Id.IntegerValue == (int)BuiltInCategory.OST_Mass)
+                return false;
             var geom = familyInstance.get_Geometry(new Options());
             return geom.OfType<Autodesk.Revit.DB.GeometryInstance>()
                 .Any(x => x.GetInstanceGeometry().Any());