+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Improve a Systemtests, CanPurgeUnusedElementsFromDocument, which use lots of element id and will fail due to changes in RevitAPI.
+* Fix a issue that it will have an offset when select a face or faces from Mass FamilyInstance into Dynamo.
 
 ## 0.2.21
 * Restore the Hide Dimension nodes and Add Dimension.ByElementDirection